spec(021): GREEN — add public/.assetsignore + bun:test gate
