+++ public/.assetsignore
@@ -1,0 +1,2 @@
+_worker.js
+_worker.js/**
