+++ specs/active/021-assetsignore-worker/tasks.md
@@ -1,10 +1,10 @@
 # Tasks
 
-- [ ] 1. Author `scripts/smoke-assetsignore.ts` with full assertion logic (existence + exact content match + clear diagnostics).
+- [ ] 1. Author `scripts/smoke-assetsignore.test.ts` — dual-mode (bun:test + script) assertion of existence + exact content of `public/.assetsignore`, with clear diagnostics on miss.
   - agent: main
   - depends: []
-  - file_targets: [scripts/smoke-assetsignore.ts]
-  - boundary: [scripts/smoke-assetsignore.ts]
+  - file_targets: [scripts/smoke-assetsignore.test.ts]
+  - boundary: [scripts/smoke-assetsignore.test.ts]
 - [ ] 2. Create `public/.assetsignore` containing exactly `_worker.js\n_worker.js/**\n` (two lines, trailing newline).
   - agent: main
   - depends: [1]
